Move helper function to 'utils' and 'MongoDBService' type to own file.
