+++ packages/realm/src/app-services/MongoDBService.ts
@@ -1,0 +1,32 @@
+////////////////////////////////////////////////////////////////////////////
+//
+// Copyright 2023 Realm Inc.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+// http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+//
+////////////////////////////////////////////////////////////////////////////
+
+import { Document } from "bson";
+
+import { MongoDBCollection } from "../internal";
+
+/**
+ * A remote MongoDB Service enabling access to a MongoDB Atlas cluster.
+ */
+export type MongoDBService = {
+  serviceName: string;
+  db(dbName: string): {
+    name: string;
+    collection<T extends Document>(collectionName: string): MongoDBCollection<T>;
+  };
+};
